fix: address Bugbot review on computer-use integration
Three issues from the Bugbot review on commit e1ba8ed:

1. extendInfo array → map (high). Switch the YAML from sequence-of-
   single-key-mappings to a proper map. Also covers the pre-existing
   NSCameraUsageDescription / NSMicrophoneUsageDescription entries that
   had the same shape.

2. Sync codesign (medium). seedComputerUseHelper used execFile with a
   callback (fire-and-forget), so the function returned before signing
   completed. Switch to execFileSync and log only after signing
   finishes — bridges can be spawned the moment the function returns.

3. Lazy-load the pi-computer-use package (low). setup.ts statically
   imported seedComputerUseHelper from computer-use-tool.ts, which
   eagerly pulled in the entire package and its Swift bridge runtime at
   app start. Move the seed helper into a new computer-use-helper.ts
   that has no package import; setup.ts statically imports the helper
   (cheap) and continues to dynamically import the extension factory.
