+++ apps/desktop/electron-builder.yml
@@ -19,10 +19,10 @@ mac:
   hardenedRuntime: true
   notarize: true
   extendInfo:
-    - NSCameraUsageDescription: Application requests access to the device's camera.
-    - NSMicrophoneUsageDescription: Application requests access to the device's microphone.
-    - NSAppleEventsUsageDescription: Inteligir uses Apple Events to control other applications when you ask the agent to act on them.
-    - NSScreenCaptureUsageDescription: Inteligir captures the screen so the agent can see and interact with native applications.
+    NSCameraUsageDescription: Application requests access to the device's camera.
+    NSMicrophoneUsageDescription: Application requests access to the device's microphone.
+    NSAppleEventsUsageDescription: Inteligir uses Apple Events to control other applications when you ask the agent to act on them.
+    NSScreenCaptureUsageDescription: Inteligir captures the screen so the agent can see and interact with native applications.
   extraResources:
     - from: ../../node_modules/@injaneity/pi-computer-use/prebuilt/macos/${arch}
       to: computer-use/${arch}
